revert: remove PR #106 from main

+++ src/community_bot/application/economy.py
@@ -19,7 +19,6 @@ if TYPE_CHECKING:
     import datetime
     from collections.abc import Mapping, Sequence
     from contextlib import AbstractAsyncContextManager
-    from decimal import Decimal
     from pathlib import Path
     from uuid import UUID
 
@@ -83,8 +82,8 @@ class LedgerHistoryItem:
     transaction_id: UUID
     member_id: UUID
     transaction_type: str
-    credit_delta: Decimal
-    experience_delta: Decimal
+    credit_delta: int
+    experience_delta: int
     comment: str | None
     created_at: datetime.datetime
 
@@ -102,10 +101,10 @@ class ReconciliationMismatch:
     """One member whose caches differ from immutable ledger sums."""
 
     member_id: UUID
-    expected_credit_balance: Decimal
-    actual_credit_balance: Decimal
-    expected_experience_total: Decimal
-    actual_experience_total: Decimal
+    expected_credit_balance: int
+    actual_credit_balance: int
+    expected_experience_total: int
+    actual_experience_total: int
 
 
 class EconomyMutationPort(Protocol):
